ui(tx-applet): relocate VOX row from PhoneCwApplet to TxApplet (3M-3a-iii bench polish)
Bench feedback (2026-05-04): operators want the VOX toggle next to MOX
and TUNE on the TxApplet right pane (where they engage TX), not
buried on the Phone tab of PhoneCwApplet.  The full VOX row including
the live peak strip, threshold slider, and Hold slider relocates as a
unit so operators retain the bench-tuning surface in the same place
they engage VOX.

Layout: TxApplet right pane gains a new row 4b underneath TUNE/MOX:
  [VOX] [Threshold===] -N dB [Hold===] N ms
        [thin DexpPeakMeter strip with red threshold marker]
PhoneCwApplet Phone tab loses Control #10 (VOX row) entirely.  DEXP
row (Control #11) stays on PhoneCwApplet - only VOX moves.

Wiring unchanged: TM voxEnabled <-> Mox setVoxEnabled (gated by voice
mode) -> voxRunRequested -> TxChannel setVoxRun -> SetDEXPRunVox.  The
keying chain (pushvox -> onVoxActive -> setMox) is untouched.

100 ms VOX peak-meter poller moves to TxApplet.  Continuous (not
MOX-gated) since GetDEXPPeakSignal is the live DEXP detector
envelope, not the TX-pipeline meter.  PhoneCwApplet's pollDexpMeters
shrinks to just the DEXP peak update.

Right-click on TxApplet VOX button still opens Setup -> Transmit ->
DEXP/VOX (same signal pattern as PhoneCwApplet's DEXP button).

Tests:
- tst_tx_applet_vox_relocated covers the new TxApplet bindings + right-
  click signal + 100 ms peak-meter timer + DexpPeakMeter presence
- tst_phone_cw_applet_vox_dexp_wiring renamed to
  tst_phone_cw_applet_dexp_wiring (DEXP-only scope) + a new case pinning
  that the VOX object names are absent on PhoneCwApplet after the move
